common_layer/requirements.txt
elasticsearch>=8.6.0,<9.0.0
boto3==1.36.5
python-dotenv==1.0.1